posts: make /posts?random=1 redirect to /posts?tags=order:random
This is a breaking change because it means the implicit `order:random`
tag added by the `random=1` param will now be counted against the user's
tag limit, so two-tag random searches will no longer work for Members.

(Also it's kind of broken if the user already has another order: metatag
in the search. Solution: don't do that.)